+++ src/main/java/dev/ultimatchamp/bettergrass/mixin/SectionBuilderMixin.java
@@ -3,9 +3,10 @@ package dev.ultimatchamp.bettergrass.mixin;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import com.mojang.blaze3d.systems.VertexSorter;
-import dev.ultimatchamp.bettergrass.BetterGrassifyBakedModel;
+import dev.ultimatchamp.bettergrass.model.BetterGrassifyBakedModel;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
+import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.BufferBuilder;
 import net.minecraft.client.render.block.BlockRenderManager;
 import net.minecraft.client.render.chunk.BlockBufferAllocatorStorage;
@@ -41,22 +42,32 @@ public class SectionBuilderMixin {
                         @Local MatrixStack matrixStack,
                         @Local BufferBuilder bufferBuilder,
                         @Local Random random) {
+        var hasSnowNeighbour = BetterGrassifyBakedModel.hasSnowNeighbour(renderRegion, blockPos.up());
 
-        if (BetterGrassifyBakedModel.canHaveSnowLayer(blockPos.up())) {
-            matrixStack.push();
-            matrixStack.translate(0, 1, 0);
-            blockRenderManager.renderBlock(Blocks.SNOW.getDefaultState(), blockPos.up(), renderRegion, matrixStack, bufferBuilder, true, random);
-            matrixStack.pop();
+        if (hasSnowNeighbour) {
+            if (BetterGrassifyBakedModel.canHaveSnowLayer(renderRegion, blockPos.up())) {
+                matrixStack.push();
+                matrixStack.translate(0, 1, 0);
+                blockRenderManager.renderBlock(Blocks.SNOW.getDefaultState(), blockPos.up(), renderRegion, matrixStack, bufferBuilder, true, random);
+                matrixStack.pop();
+            }
         }
     }
 
     @ModifyVariable(method = "build", at = @At("STORE"), ordinal = 0)
     private BlockState setGrassState(BlockState state, @Local(ordinal = 2) BlockPos blockPos) {
-        if (BetterGrassifyBakedModel.canHaveSnowLayer(blockPos.up())) {
-            return state.with(Properties.SNOWY, true);
-        } else {
-            return state;
+        var world = MinecraftClient.getInstance().world;
+
+        if (world != null) {
+            var hasSnowNeighbour = BetterGrassifyBakedModel.hasSnowNeighbour(world, blockPos.up());
+
+            if (hasSnowNeighbour) {
+                if (BetterGrassifyBakedModel.canHaveSnowLayer(world, blockPos.up())) {
+                    return state.with(Properties.SNOWY, true);
+                }
+            }
         }
+        return state;
     }
 }
 //?}